Update Gradle version to 8.9
Gradle version has been updated to 8.9 in gradle-wrapper.properties. Along with this, corresponding changes to point to the updated version have been made in .github/workflows/gradle.yml and build.gradle.kts. Also, minor adjustments to the gradlew and gradlew.bat scripts have been done as part of this update.